<new>:"Реализована новая имплементация хранилища, использующая базу данных. Основные методы и эндпоинты работают стабильно."

+++ src/main/java/ru/yandex/practicum/filmorate/exception/FilmAlreadyExistException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exception;
 
 public class FilmAlreadyExistException extends RuntimeException {
-    public FilmAlreadyExistException(int id) {
+    public FilmAlreadyExistException(long id) {
         super(String.format("Film with id = %s already exist.", id));
     }
 }
