+++ src/main/java/ru/yandex/practicum/filmorate/exception/UserAlreadyExistException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exception;
 
 public class UserAlreadyExistException extends RuntimeException {
-    public UserAlreadyExistException(int id) {
+    public UserAlreadyExistException(long id) {
         super(String.format("User with id = %s already exist.", id));
     }
 }
